Added tweets, times, and styling
Page now shows all tweets available to the user, as well as the time
at which they were posted. Styling was also added via the CSS stylesheet.
A page title was added for clarity, as well as a list of people the
user is following, for easier navigation of the site.