Add filtering by twittler user
User can now filter the tweets in their feed by user. This can be
done by clicking on a username in the tweet feed, or by selecting
from the list of people following on the left hand side. To get back
to the homepage, where all tweets can be viewed, the page title/header
can be clicked.