Got moment.js working for user friendly time view
The time at which a tweet was posted now updates on the page every
minute. And when a new tweet is added, the time at which it is posted
is tracked, and again updated every minute. Still able to switch between
different users and view feeds. Most recent tweets sit at the top of the
page.